+++ Sources/HIGThemesContract/HIGTheme+Defaults.swift
@@ -36,6 +36,7 @@ extension HIGTheme {
     public var modal: any HIGModalTokens { HIGSystemModalTokens() }
     public var tooltip: any HIGTooltipTokens { HIGSystemTooltipTokens() }
     public var popover: any HIGPopoverTokens { HIGSystemPopoverTokens() }
+    public var drawer: any HIGDrawerTokens { HIGSystemDrawerTokens() }
     public var progress: any HIGProgressTokens { HIGSystemProgressTokens() }
     public var alert: any HIGAlertTokens { HIGSystemAlertTokens() }
     public var toast: any HIGToastTokens { HIGSystemToastTokens() }